requirements.txt
# Core Framework
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-multipart>=0.0.20
pydantic-settings>=2.7.0

# LangChain Ecosystem (v1.x)
langchain>=1.2.0
langchain-core>=1.2.0
langchain-community>=0.4.0
langchain-openai>=1.1.0
langchain-experimental>=0.4.0
langchain-chroma>=1.1.0
langchain-text-splitters>=0.3.8

# Vector Stores
chromadb>=1.0.0
langchain-pinecone>=0.2.13
pinecone>=5.4.0

# Retrieval
rank-bm25>=0.2.2

# Reranking
cohere>=5.20.0

# Document Loading
pypdf>=5.1.0

# Web Search
tavily-python>=0.7.0

# Database
sqlalchemy>=2.0.0

# Evaluation
ragas>=0.4.0
datasets>=3.2.0

# Frontend
streamlit>=1.41.0

# Utilities
python-dotenv>=1.0.0
httpx>=0.28.0
tenacity>=9.0.0
